strands-agents/waf-logs-in-clickhouse-with-mcp/requirements.txt
clickhouse-connect>=0.8.16
pip-system-certs>=4.0
json
strands-agents
strands-agents-tools